refactor(http-py): move remaining Python bindings from http crate (#852)
Move the remaining PyO3 binding modules out of dcc-mcp-http and into
 the dedicated dcc-mcp-http-py crate:

- config (McpHttpConfig Python wrapper)
- readiness (ReadinessProbe)
- server (McpServerHandle)
- skill_server (McpHttpServer)
- bridge (bridge registry helpers and factory functions)
- output_dynamic (ToolSpec / OutputCapture)
- resources_handle and prompts_handle

The root _core extension already registers through dcc-mcp-http-py, so
this completes the module-level Python boundary move without changing the
Python API. dcc-mcp-http now exposes the Rust HTTP application surface and
small public helpers needed by the py crate, while Python registration and
classes live in dcc-mcp-http-py.

Implementation notes:
- Promote LiveMeta and build_resource_registry from crate-private to public
  HTTP application helpers so dcc-mcp-http-py can build the same server state
  that the Rust path uses.
- Disable doctests for dcc-mcp-http-py because its public API docs contain
  Python examples, not Rust snippets.
- Add auto-initialize for dcc-mcp-http-py tests so moved PyO3 resource-handle
  tests can run in the new crate.

Verification:
- cargo check --workspace
- cargo check -p dcc-mcp-http-py --features python-bindings
- cargo check -p dcc-mcp-core --features python-bindings
- cargo test -p dcc-mcp-http-py --features python-bindings
- cargo clippy -p dcc-mcp-http-py --all-targets -- -D warnings
