+++ crates/dcc-mcp-http-py/src/readiness.rs
